added presentation exercise and refactored seed
